+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. This projec
 - Add laravel 5.3 support to composer.json
 - Change $this->currentRoute to $this->currentRouteName
 - Remove compareDotArrays() completely and use laravel str_is
+- Change function param $route to $routePattern
 
 ## v1.0.6